registration closed

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -52,9 +52,18 @@ const Mutations = {
       `{ 
       id
       title
+      user {id}
     }`
     );
     //2. Check if they own the item or have permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
 
     //3. Delete it!!
     return ctx.db.mutation.deleteItem({ where }, info);
